new update many file

+++ lib/PR/screen/download/resource/pull_customer.dart
@@ -1,14 +1,18 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:mimos/PR/dao/customer_pr_dao.dart';
+import 'package:mimos/PR/dao/customer_wsp_dao.dart';
 import 'package:mimos/PR/model/default/download_model.dart';
 import 'package:mimos/PR/model/response/list_response.dart';
+import 'package:mimos/PR/repo/customer_wsp.dart';
 import 'package:mimos/PR/repo/master_data_repo.dart';
 import 'package:mimos/helper/session_manager.dart';
 
 class PullCustomer {
   var _dao = CustomerPRDao();
   var _repo = MasterDataRepo();
+  var _daoCustWsp = CustomerWspDao();
+  var _repoCustWsp = CustomerWspRepo();
   var _model = DownloadModel();
 
   Future<DownloadModel> init() async {
@@ -26,21 +30,76 @@ class PullCustomer {
     _model = await init();
     yield _loading();
 
+    var res = await pullCustomer(date: date);
+
+    if(res.status){
+      var custWsp = await pullCustomerWsp();
+      yield await _success(res);
+    }else{
+      yield _failed(res.message);
+    }
+
+  }
+
+  // Stream<DownloadModel> pullCustomer({@required String date}) async* {
+  //   _model = await init();
+  //   yield _loading();
+  //
+  //   var response = await _repo.pullCustomerPR(
+  //     userId: await session.getUserId(),
+  //     date: date,
+  //   );
+  //
+  //   if (response.status) {
+  //     _dao.truncate();
+  //     _dao.insertAll(response.list);
+  //     print("$runtimeType: ${response.message}");
+  //
+  //     yield await _success(response);
+  //   } else {
+  //     print("$runtimeType: ${response.message}");
+  //
+  //     yield _failed(response.message);
+  //   }
+  // }
+
+  Future<ListResponse> pullCustomer({@required String date}) async {
     var response = await _repo.pullCustomerPR(
       userId: await session.getUserId(),
       date: date,
     );
 
     if (response.status) {
-      _dao.truncate();
-      _dao.insertAll(response.list);
-      print("$runtimeType: ${response.message}");
+      if (response.list != null) {
+        _dao.truncate();
+        _dao.insertAll(response.list);
+        print("$runtimeType: ${response.message}");
 
-      yield await _success(response);
+        return response;
+      } else {
+        return response;
+      }
     } else {
-      print("$runtimeType: ${response.message}");
+      return response;
+    }
+  }
 
-      yield _failed(response.message);
+  Future<ListResponse> pullCustomerWsp() async {
+    var response = await _repoCustWsp.pull(
+      customernos: await _dao.getAllCustomerno(),
+    );
+
+    if (response.status) {
+      if (response.list != null) {
+        _daoCustWsp.truncate();
+        _daoCustWsp.insertAll(response.list);
+
+        return response;
+      } else {
+        return response;
+      }
+    } else {
+      return response;
     }
   }
